REF: change fitness to double type; add control statement for index

+++ src/WeightedRandomGenerator.java
@@ -8,9 +8,9 @@ import java.util.Random;
 //Generated weighted random numbers according to the weights array.
 public class WeightedRandomGenerator {
 
-    public int nextInt(float[] totals) {
-        float sumOfWeights = totals[totals.length - 1];
-        float randomNumber = Rand.getFloat() * sumOfWeights;
+    public int nextInt(double[] totals) {
+        double sumOfWeights = totals[totals.length - 1];
+        double randomNumber = Rand.getDouble() * sumOfWeights;
         int index = Arrays.binarySearch(totals, randomNumber);
         index = (index >= 0) ? index : - index - 1;
         return index;
